fix errors

+++ ReportingService/ReportingService.Core/Dtos/TransactionDto.cs
@@ -8,7 +8,7 @@ public class TransactionDto
     public AccountDto Account { get; set; }
     public TransactionType TransactionType { get; set; }
     public CurrencyType CurrencyType { get; set; }
-    public double? Commission { get; set; }
+    public decimal? Commission { get; set; }
     public decimal Amount { get; set; }
     public DateTime Date { get; set; }
 }
